requirements.txt
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.12.2
beniget==0.4.1
bleach==6.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudpickle==2.0.0
comm==0.1.4
cycler==0.11.0
Cython==0.29.24
daal==2021.4.0
dask==2021.11.1
debugpy==1.8.0
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
exceptiongroup==1.1.3
fastjsonschema==2.18.0
filelock==3.3.2
fortran-language-server==1.12.0
fqdn==1.5.1
fsspec==2021.11.0
gast==0.5.3
greenlet==2.0.2
h5py==3.10.0
hypothesis==6.24.3
idna==3.3
importlib-metadata==6.8.0
iniconfig==1.1.1
intel-openmp==2021.4.0
ipp==2021.4.0
ipykernel==6.25.2
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.18.0
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.3.2
line-profiler==3.3.1
llvmlite==0.37.0
locket==0.2.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.4.3
matplotlib-inline==0.1.3
memory-profiler==0.58.0
mistune==3.0.1
mkl==2021.4.0
mpi4py==3.1.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numba==0.54.1
numexpr==2.7.3
numpy==1.20.3
obspy==1.4.0
overrides==7.4.0
packaging==23.1
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pybind11==2.8.1
pybind11_global==2.12.0
pycparser==2.21
Pygments==2.10.0
pyparsing==2.4.7
pyproj==3.6.1
pytest==6.2.5
python-dateutil==2.8.2
python-json-logger==2.0.7
pythran==0.10.0
pytz==2021.3
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
scikit-learn==1.0.1
scipy==1.7.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
tbb==2021.4.0
tdqm==0.0.1
terminado==0.17.1
threadpoolctl==3.0.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
widgetsnbextension==4.0.9
xgboost==1.5.0
zipp==3.16.2
